+++ src/main/java/http/helpers/MailService.java
@@ -1,19 +1,20 @@
 package http.helpers;
 
 import javax.mail.*;
-import java.util.Properties;
 
+/**
+ * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.tools.http.helpers.MailService}
+ */
+@Deprecated()
 public class MailService {
 
-    private Session session;
-    private Store store;
-    private Folder folder;
-
-    private String protocol = "imaps";
-    private MailFolder folderName = MailFolder.INBOX;
+    private final net.itarray.automotion.tools.http.helpers.MailService delegatee;
+    public MailService() {
+        delegatee = new net.itarray.automotion.tools.http.helpers.MailService();
+    }
 
     public boolean isLoggedIn() {
-        return store != null && store.isConnected();
+        return delegatee.isLoggedIn();
     }
 
     /**
@@ -21,58 +22,40 @@ public class MailService {
      */
     public void login(String host, int port, String username, String password)
             throws Exception {
-        String fName = getFolderName(folderName);
-        URLName url = new URLName(protocol, host, port, fName, username, password);
-
-        if (session == null) {
-            Properties props = null;
-            try {
-                props = System.getProperties();
-            } catch (SecurityException sex) {
-                props = new Properties();
-            }
-            session = Session.getInstance(props, null);
-        }
-        store = session.getStore(url);
-        store.connect();
-        folder = store.getFolder(url);
-
-        folder.open(Folder.READ_WRITE);
+        delegatee.login(host, port, username, password);
     }
 
     /**
      * to logout from the mail host server
      */
     public void logout() throws MessagingException {
-        folder.close(false);
-        store.close();
-        store = null;
-        session = null;
+        delegatee.logout();
     }
 
     public int getMessageCount() {
-        int messageCount = 0;
-        try {
-            messageCount = folder.getMessageCount();
-        } catch (MessagingException me) {
-            me.printStackTrace();
-        }
-        return messageCount;
+        return delegatee.getMessageCount();
     }
 
     public Message[] getMessages() throws MessagingException {
-        return folder.getMessages();
+        return delegatee.getMessages();
     }
 
     public Message getLastMesage() throws MessagingException {
-        return getMessages()[getMessageCount() - 1];
+        return delegatee.getLastMesage();
     }
 
     public MailService setFolder(MailFolder folder){
-        if (!isLoggedIn()) {
-            this.folderName = folder;
+        switch (folder) {
+            case INBOX:
+                delegatee.setFolder(net.itarray.automotion.tools.http.helpers.MailService.MailFolder.INBOX);
+                break;
+            case SPAM:
+                delegatee.setFolder(net.itarray.automotion.tools.http.helpers.MailService.MailFolder.SPAM);
+                break;
+            case TRASH:
+                delegatee.setFolder(net.itarray.automotion.tools.http.helpers.MailService.MailFolder.TRASH);
+                break;
         }
-
         return this;
     }
 
@@ -82,24 +65,4 @@ public class MailService {
         TRASH
     }
 
-    private String getFolderName(MailFolder folder){
-        String folderName;
-
-        switch (folder){
-            case INBOX:
-                folderName = "INBOX";
-                break;
-            case SPAM:
-                folderName = "[Gmail]/Spam";
-                break;
-            case TRASH:
-                folderName = "[Gmail]/Trash";
-                break;
-            default:
-                folderName = "INBOX";
-                break;
-        }
-
-        return folderName;
-    }
 }
